move to handles/resize.tsx
